Add semifinal/ch1 and contents.
The idea here is that, as I finish the first post-edit round of writing on each
chapter, I put the current state of the code here. Once that's done for all
chapters, I plan to make things cross-platform (I know, it's better to do that
earlier), and then remove the semifinal dir and relocate the ch? dirs in the
root of the repo, and clean up the entire repo. By the end, I hope the entire
repo feels relatively clean. I may consider renaming the repo based on the title
of the book and optionally pulling out an eatyguy-specific repo if I want to add
any more features to that.