Restructure lua dep in semifinal.
The goal of this is to avoid having a precompiled library file in this repo.
Instead, if a user is working on macOS or Linux, the makefile can build liblua.a
on the fly as needed. I was considering the alternative of bundling in prebuilt
library files, but then I'd have to anticipate many different systems; this way,
at least theoretically, anyone who can compile C can get a working liblua.a
file. Of course, the Windows setup is different.